fix：すべてがmain.tsのをやめた｜fix:gitginoreを修正した
