+++ .gitignore
@@ -1,3 +1,5 @@
 dist/
 node_modules/
 .eslintrc.json
+.env
+settings.json
